put tests in v2 and v3 packages
